Extract service-connectivity validators from validations.py
Third and final thematic extraction from modules/validations.py:

- modules/validations_services.py (~590 lines, new):
  18 functions covering external-service connectivity validation.
  Every validator follows the same shape: pull credentials from a
  JSON payload, attempt a real network round-trip, return a Flask
  jsonify response.  Organised by service family:
  * Plex family: validate_plex_server, validate_tautulli_server,
    validate_trakt_server
  * Notification services: validate_gotify_server,
    validate_ntfy_server, validate_apprise_server,
    validate_webhook_server, validate_notifiarr_server
  * Arr apps: validate_radarr_server/payload,
    validate_sonarr_server/payload
  * Metadata/rating APIs: validate_omdb_server,
    validate_github_server, validate_tmdb_server,
    validate_mdblist_server
  * Anime tracker: validate_mal_server
  Plus the shared URL pre-flight helper _validate_service_url.

modules/validations.py: 596 -> 143 (-453 / -76.0%).

## Sprint 4 combined validations.py totals

1828 (start) -> 143 (now) = -1686 lines / -92.2% over 3 PRs
(#1519 overlay-images, #1520 yaml-files, this PR).

modules/validations.py is now essentially a backward-compatibility
facade: re-export blocks + 2 tiny ISO code helpers that don't fit
any other cluster.  Added a module docstring explaining the split.

## Test fix

One test (test_validate_plex_fetches_sections_once) monkey-patched
qs_module.validations.PlexServer expecting the function under test
to look up PlexServer via validations.__dict__.  After the extraction
validate_plex_server lives in modules.validations_services and looks
up PlexServer in ITS OWN namespace, so the patch didn't propagate.
Updated the test to patch modules.validations_services.PlexServer
directly.  This is the only test change in the Sprint 4 refactor;
every other backward-compat pattern (requests.get, persistence.retrieve_settings)
works via Python module identity since those are shared MODULE objects.

## Load-bearing imports retained in validations.py

- requests (14 test-suite monkeypatches)
- persistence (1 test-suite monkeypatch)

Both marked with noqa: F401 comments explaining why.

AST-verified: all 18 moved functions are byte-for-byte identical to
origin/develop, including validate_ntfy_server (which contains a
non-ASCII bell character in a test message).
